requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

streamlit==1.33.0
python-dotenv==1.0.1
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.4
langchain-huggingface==0.0.3
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4
